splatnew: ensure performance of object users
Types which permit nargs < nfields in their constructor are required by correctness
to take a notable performance hit. Since right now this is only used by NamedTuple,
and it already expects to enforce this condition, it seems like a bad idea to make all future users slow
for a feature that is not needed.